requirements.txt
torch>=2.0.0
numpy>=1.24.0
tokenizers>=0.15.0
datasets>=2.16.0
transformers>=4.36.0
tqdm>=4.66.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
pyyaml>=6.0
